feat: move preprocessing.c fucntions to data_init.c then deleted preprocess.c
